requirements.txt
requests
pandas
sqlalchemy
streamlit
python-dotenv
psycopg2